server/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
fastapi==0.110.0
grpcio==1.62.1
grpcio-tools==1.62.1
idna==3.6
protobuf==4.25.3
pydantic==2.6.4
pydantic_core==2.16.3
setuptools==68.2.2
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
wheel==0.41.2
